job_crawler_project/requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
psycopg2-binary==2.9.10